requirements.txt
absl-py==0.15.0
anyio==3.2.1
appnope==0.1.2
APScheduler==3.6.3
argon2-cffi==20.1.0
astroid==2.8.0
astunparse==1.6.3
async-generator==1.10
attrs==21.2.0
Automat==20.2.0
Babel==2.9.1
backcall==0.2.0
bleach==3.3.0
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
clang==5.0
click==8.0.1
coinbase==2.1.0
constantly==15.1.0
construct==2.10.67
cssselect==1.1.0
cycler==0.10.0
decorator==5.0.9
defusedxml==0.7.1
dnspython==2.2.0
docopt==0.6.2
entrypoints==0.3
Flask==2.0.1
flatbuffers==1.12
gast==0.4.0
geographiclib==1.52
geopy==2.2.0
google-api-core==2.0.1
google-api-python-client==2.21.0
google-auth==2.1.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
googleapis-common-protos==1.53.0
grpcio==1.41.1
gspread==3.6.0
gspread-dataframe==3.2.1
h5py==3.1.0
httplib2==0.18.1
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
ipykernel==5.5.5
ipython==7.24.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
isort==5.9.3
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
json5==0.9.6
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyter-server==1.8.0
jupyterlab==3.0.16
jupyterlab-pygments==0.1.2
jupyterlab-server==2.6.0
jupyterlab-widgets==1.0.2
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
lazy-object-proxy==1.6.0
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.2
mccabe==0.6.1
mistune==0.8.4
multitasking==0.0.9
nbclassic==0.3.1
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
numpy==1.19.5
oauth2client==4.1.3
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.5
pandoc==1.1.0
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
pipreqs==0.4.10
platformdirs==2.3.0
plumbum==1.7.0
ply==3.11
prometheus-client==0.11.0
prompt-toolkit==3.0.19
protobuf==3.18.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.9.9
pycryptodomex==3.11.0
Pygments==2.9.0
pykerberos==1.2.1
pylint==2.11.1
pymongo==4.0.1
pyparsing==2.4.7
pyquery==1.4.3
pyrsistent==0.17.3
pysmi==0.3.4
pysnmp==4.4.12
pyTelegramBotAPI==3.8.1
python-dateutil==2.8.1
python-dotenv==0.18.0
python-snappy==0.6.0
python-telegram-bot==13.9
pytz==2021.1
PyYAML==6.0
pyzmq==22.1.0
qgrid==1.3.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
Send2Trash==1.7.1
six==1.15.0
sniffio==1.2.0
sportsipy==0.6.0
tensor==0.3.6
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.7.0
termcolor==1.1.0
terminado==0.10.1
testpath==0.5.0
tinydb==4.6.1
toml==0.10.2
tornado==6.1
traitlets==5.0.5
Twisted==21.7.0
typing-extensions==3.7.4.3
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.26.2
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.1.0
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
yarg==0.1.9
yfinance==0.1.59
zope.interface==5.4.0
